+++ src/main/java/org/javlo/actions/ListerException.java
@@ -1,0 +1,5 @@
+package org.javlo.actions;
+
+public class ListerException extends Exception {
+
+}